Cleanup...

+++ PostfixCodeCompletionTests/TestUtils/TestFile.cs
@@ -1,20 +1,17 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+using System.IO;
 using System.Reflection;
 using System.Text;
-using System.IO;
 
-namespace ASCompletion.TestUtils
+namespace PostfixCodeCompletion.TestUtils
 {
-    class TestFile : IDisposable
+    internal class TestFile : IDisposable
     {
         public string ResourceFile { get; private set; }
 
-        public string DestinationFile { get; private set; }
+        public string DestinationFile { get; }
 
-        public TestFile(string resourceFile)
-            : this(resourceFile, GetTempFileName(resourceFile))
+        public TestFile(string resourceFile) : this(resourceFile, GetTempFileName(resourceFile))
         {
         }
 
@@ -22,21 +19,18 @@ namespace ASCompletion.TestUtils
         {
             DestinationFile = destinationFile;
             ResourceFile = resourceFile;
-
             File.WriteAllBytes(destinationFile, ReadAllBytes(resourceFile));
         }
 
         private static string GetTempFileName(string baseFileName)
         {
-            string temp = Path.GetTempFileName();
-
+            var temp = Path.GetTempFileName();
             return Path.GetFileNameWithoutExtension(temp) + Path.GetExtension(baseFileName);
         }
 
         public void Dispose()
         {
-            if (File.Exists(DestinationFile))
-                File.Delete(DestinationFile);
+            if (File.Exists(DestinationFile)) File.Delete(DestinationFile);
         }
 
         public static string ReadAllText(string resourceFile)
@@ -53,10 +47,9 @@ namespace ASCompletion.TestUtils
         public static byte[] ReadAllBytes(string resourceFile)
         {
             var asm = Assembly.GetExecutingAssembly();
-
             using (var stream = asm.GetManifestResourceStream(resourceFile))
             {
-                byte[] buffer = new byte[stream.Length];
+                var buffer = new byte[stream.Length];
                 stream.Read(buffer, 0, buffer.Length);
                 return buffer;
             }
